requirements.txt
annotated-types==0.7.0
anyio==4.11.0
archspec @ file:///home/conda/feedstock_root/build_artifacts/archspec_1737352602016/work
attrs==25.4.0
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1749686179973/work
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1756599213019/work
cachetools==6.2.2
cattrs==25.3.0
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1759648874697/work/certifi
cffi @ file:///Users/runner/miniforge3/conda-bld/cffi_1761202952988/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1760437218288/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
conda @ file:///Users/runner/miniforge3/conda-bld/conda_1760108736307/work/conda-src
conda-libmamba-solver @ file:///home/conda/feedstock_root/build_artifacts/conda-libmamba-solver_1745834476052/work/src
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1736345463896/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1751548120229/work
cryptography==46.0.3
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1734729835256/work
flatbuffers==25.9.23
frozendict @ file:///Users/runner/miniforge3/conda-bld/frozendict_1756048002891/work
google-api-core==2.28.1
googleapis-common-protos==1.72.0
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1760286409563/work
jh2==5.0.10
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1733814567314/work
jsonpointer @ file:///Users/runner/miniforge3/conda-bld/jsonpointer_1756754188614/work
libmambapy @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_libmambapy_1760967438/work/libmambapy
linkify-it-py==2.0.3
markdown-it-py==4.0.0
mdit-py-plugins==0.5.0
mdurl==0.1.2
menuinst @ file:///Users/runner/miniforge3/conda-bld/menuinst_1761299914766/work
niquests==3.15.2
numpy==2.3.5
oauthlib==3.3.1
ollama==0.6.1
openmeteo_requests==1.7.4
openmeteo_sdk==1.23.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas==2.3.3
pillow==12.0.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1759953252/work
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1747339660894/work
proto-plus==1.26.1
protobuf==6.33.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycosat @ file:///Users/runner/miniforge3/conda-bld/pycosat_1757744682161/work
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pydantic==2.12.5
pydantic_core==2.41.5
PyGithub==2.8.1
Pygments==2.19.2
PyJWT==2.10.1
PyNaCl==1.6.1
pyparsing==3.2.5
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
qh3==1.5.6
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1755614211359/work
requests-cache==1.2.1
requests-oauthlib==2.0.0
retry-requests==2.0.0
rich==14.2.0
rsa==4.9.1
ruamel.yaml @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml_1761160761767/work
ruamel.yaml.clib @ file:///Users/runner/miniforge3/conda-bld/ruamel.yaml.clib_1760564263846/work
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
textual==6.6.0
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
truststore @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_truststore_1753886790/work
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
uritemplate==4.2.0
url-normalize==2.2.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
urllib3-future==2.14.907
wassima==2.0.2
watchdog==6.0.0
wheel==0.45.1
zstandard==0.25.0
